Move SpotifyService
Moved SpotifyService to backend and corrected the spelling of suscribeToChanges. Fixed line 45 where the onClick listener was connected to the pauseSnippet instead of the playSnippet.
